Make SharedMember(shared_ptr) constructors public
It seems a bit paranoid to restrict creation of SharedMember wrappers
around shared_ptr: SharedMember is really just some convenience
functions for Members added to shared_ptr.

This makes the constructors public and implicit, so that any shared_ptr
is implicitly convertible to a shared_ptr, and adds a shared_ptr
conversion operator so that SharedMembers are also convertible back to
shared_ptrs.