Use boost::shared_lock to fix recursive insert bug
Prior to this commit, code such as this was not working as expected:

    class A : Member {
        ...
        virtual void added() override {
            simulation()->spawn<B>();
        }
    }
    class B : Member { ... }
    sim->spawn<A>();

Rather than have two members created, the "A" member would get inserted,
but the "B" member wouldn't, because at the time of the inner spawn<B>
call, the runLock was held by the outer spawn<A> call.  As a
consequence, the B insertion would be added to the deferred insert
queue.

The deferred insert queue, however, is only processed at the end of
each stage, so the very next stage wouldn't actually have the "B"
member as part of the simulation.

Worse, there was deadlock potential here: if sim->spawn<A>() was itself
called during some run stage, it would be correctly deferred, but then
the deferral queue would be locked while the queue was processed.
Attempting to insert A would then result in an attempt to insert B,
which would be (rightly) sent to the deferral queue: but locking the
deferral queue would deadlock, because it's still being processed.

This commit changes the runLock()/runLockTry() methods to use
boost::shared_lock/boost::shared_mutex for the run lock: this means that
recursive adding is fine: it just obtains multiple (shared) locks.  (A
recursive_mutex would not be fine here: the lock is not only locking out
other threads, it's also designed to lock out the same thread during a
run).

It also changes the deferred queue processing to pull one queue element
at a time, then unlocking, then processing the actual insert/removal;
thus a recursive add attempt (like the above) can properly add to the
end of the deferred queue.  (It has to be at the end, rather than
processing immediately, because we preserve insertion order).

This does, however, add a dependency on boost_thread.