Good::Continuous -> Good; added atom()
Good::Continuous is a bit cumbersome: basically all goods should be
continuous unless otherwise wanted.  This repurposes Good to be a
continuous good, making Good::Continuous a deprecated typedef for Good.

To better allow discrete good use, this adds an atom() method that
returns the smallest increment--0.0 for continuous, and 1.0 (by
default) for discrete.

It also moves Good::Discrete to good::Discrete with its own header in
eris/good/Discrete.hpp.  This may break things, but as far as I can
tell, creativity was the only place it was used (and even then, it
isn't technically forcing discreteness, it's just a notional base
class).