De-virtualize several methods; remove default canProduceAny; doc fixes
A lot of the methods in Firm are virtual but really don't need to be,
especially since they are simply complexity-handling interfaces that
themselves call virtual methods to do the firm-specific things.

Also remove default canProduceAny implementation that returns 0: that's
a useless default for a firm with production, and a firm without
production should be using FirmNoProd as a base class instead.

Some of the documentation was referring to long-changed method names.