Get rid of Firm/Market::transfer()/release()
Moved the transfer/release code into the reservation methods instead,
which seems a more natural place for it and simplifies the Firm/Market
interfaces.

Updated constructor to only release if the firm/market shared pointer
is set (it won't be if the destructor is called the
uninitialized-but-valid remaining instance after being passed to the
move constructor).