Fixed Reservation move constructors
The default move constructors attempted to move the const SharedMembers,
which ends up invoking the copy constructors rather than the move
constructors.  This in turn means that firm/market remains set, and so
the destructor runs when it shouldn't.

This changes the code to use an explicit const_cast on the moved const
members so as to properly move the SharedMember away.

It also adds explicit default move/copy constructors/assignment
operators to SharedMember.