Positional<T> constructors only participate when T(Args...) is valid
Positional<T> and WrappedPositional<T> constructors were participating
whenever the Positional arguments matched; if the arguments forwarded
to the T constructor weren't valid, this showed up as a compilation
error in the Positional<T> constructor.  With this commit, the
constructors only participate when T is constructible with (Args...).

This was a problem before, which caused the
WrappedPositional<T>(Position, ...) constructor to not work, because it
would be called even when invoked with WrappedPositional<T>(Pos, Pos,
Pos, ...), then failed because T(Pos, Pos, ...) isn't a valid
constructor.  This was worked around in e5b1f8b5, which just removed the
constructor.

This commit has a better fix that: since T(Pos, Pos, ...) isn't a valid
constructor, WP<T>(Pos, ...) won't be called when the first three
arguments are three Positions, and thus it doesn't end up masking the
desired constructor.

This same issue affected Positional<T>, which has the same constructor
format, but apparently wasn't caught (likely because Positional isn't
used as much as WrappedPositional).