Filter now takes SharedMember<M>
Takes a raw const M& was a little weird, since the API takes great
pains to avoid ever having an unwrapped M anywhere else; this makes the
filters take a SharedMember instead.

This also simplifies the filter logic: the dynamic_cast in
genericFilter when not doing class filtering wasn't needed, since the
only time we get there is when T == B, and so casting a B* to a T*
isn't actually doing anything.