Allow Simulation inheritance
This allows Simulation to be used via inheritance as an alternative to
composition.  This, in particular, allows members to always access
their owning simulation, while without this they had both the
simulation (via the internal eris weak_ptr) *and* some sort of local
reference to the Simulation-containing class (for simulation-specific
settings).

Allowing this via inheritance makes life easy: members can now get to
the simulation-specific bits by casting the simulation to the local
subclass name.