Member ids: global, permanent; no implicit conversion
This changes several things related to the handling of member ids.
This is an API-breaking change, but cleans up some of the eris_id_t
handling.  Making SharedMember and Member implicit convertible to
eris_id_t everywhere was a mistake; this commit corrects it by
introducing an implicit-converting object in those places where
accepting either eris_id_t or a Member is desirable, without leaving
implicit conversion to integer semantics in place everywhere else.

Details:

Member and SharedMember<T> are no longer implicitly convertible
to eris_id_t.

Assigned member id()'s are now permanent and globally unique (rather
than just within-Simulation unique), and are assigned when the Member is
constructed and never unassigned.

SharedMember<T> is now *explicitly* convertible to bool, but the meaning
has changed somewhat: it no longer means "in a simulation" but now just
"has a Member", like std::shared_ptr.  Thus shared members without a
simulation used to be false in boolean context (which was via the
eris_id_t) but are now true.

SharedMember equality and inequality now work differently: since member
ids are never unassigned and are global equality between shared members
now works as expected across in- and out-of-simulation members.
Previously `a == a` was false when `a` was not in a simulation.
Equality is newly also true for null shared members: that is,
`SharedMember<Member>() == SharedMember<Member>()` is true when both are
nullptr members.  If just one is null, the comparison is false;
otherwise the comparison comes down to id() values.  `operator <` now
treats all nullptr members as if they had an id() of 0 (real id() values
start at 1).

To limit some of the compatibility annoyance, various methods like
remove(), dependsOn(), and many of the methods in Bundle now accept a
MemberID object (defined in types.hpp) instead of an eris_id_t that is
implicitly convertible to and from an eris_id_t, and from anything with
a direct or indirect id() method.

Since this changes breaks some compatibility already, at the same time
this also removes the second argument (the old id) from
Member::weakDepRemoved() since the concept of an "old" id no longer
exists.  (The member keeps its id even if removed and/or readded to a
simulation).